Migrate to Flatpak
Made changes mostly based on GNOME Builder's generated template, with
some tweaks to use Relm4 and to remove things I'm not currently using,
to simplify this patch.
The template uses Meson instead of Cargo for scripting builds.
`build.rs` is no more.

The header bar "devel" style is now defined by a constant `BUILD_TYPE`
which Meson writes to before building.
Just pass `-Dbuildtype=release` to Meson to remove this style.
I just temporarily change this in the manifest and rebuild the app.

The template came with `gettext` support, which I'm not inclined to use.
I'm thinking on using other localization framework else in the future.
I failed to come up with a simple way to validate the MetaInfo and
desktop files, so I left `data/meson.build` using `i18n.merge_file()`.

I wish I'd called the manifest `manifest.json`, but Builder wouldn't
recognize it.

Also changes the app icon to the one the template uses, which is a blue
diamond with gears.
The previous one was just to have something as an icon.
This new icon looks way better, even as a placeholder.