Load pets on app launch
I'm assuming if the loading failed, it was because the file didn't exist.
This is what happens when the user opens the app for the first time.
That's why I don't handle or inform anything.

However, if it failed for another reason, I don't know what to do.
The app assumes the first scenario.