Indicate current view as header bar title
Previously, the header bar said "Pet Details" even if you were editing a pet.

Since these views already have names because they're in a `gtk::Stack`, I update the title accordingly.
Now it either says "Pet Details" or "Edit Pet", depending on the current view.

I introduced `current_view` to keep track of, well, the current view :P
It's easier to watch for view changes to update the title this way.
This made the `is_in_edit_mode` field a paraphrase and could lead to inconsistencies.
I turned it into a function to preserve the readability in the other places it was being read.