Move code to show the "About" window outside of `app`
That way it were, `app` had the responsability to show modal windows and also deal with `app` things.

While it makes sense for the `app` to show modal windows, in reality the code became really messy.
Its update function was just too big.
On top of that, I almost never looked at the "Show _ Window" parts.
In fact, I _avoided_ them, because they were really close to the other code I care about.

It turns out it's possible to show modal windows without sending inputs to components.
I moved the code to present them to the `modals` module, under a new `show_window(_)` function.
This cleans `app`'s update function a lot.

This improvements aren't that big as it could be, because I removed most of the modal windows.
However, once I re-implement them, things will continue to be tidy, not messing with `app` anymore.