Add `rustfmt.toml` and enforce it
Effectively, uses tabs for indentation instead of spaces.
I'm aware this isn't the convention, but I strongly disagree with it.
While not recommended by the default style, it's also not forbidden.

From:
https://doc.rust-lang.org/beta/style-guide/index.html#the-default-rust-style

> This should not be interpreted as forbidding developers from following a non-default style, or forbidding tools from adding any particular configuration options.

Why bother?
While in Rust specifically this have little impact _for me_, for other languages this impacts me more.
Instead of using tabs for those files and spaces for others, I'll use tabs for all of them.
This includes Rust files.