Add more fields to "Add Pet" view
This view only had a field for the name.
That means added pets only had custom names; everything else were default values.

I put the remaining fields.

Previously, a pet was only added after pressing "Enter" on the keyboard, after setting a name.
Since there are more fields now, it doesn't make sense to keep this behaviour, specially because the name field is the first one.
With that in mind, I also added an "Apply" button.