Move "No Pet Selected" into the stack
Currently, the first view is "Pet Details".
`content` decides whether to show it or "No Pet Details", based on the optional `selected_pet`.
The stack currently holds only "Pet Details" and "Edit Pet" views.
The "No Pet Selected" and "Add Pet" views are shown based on other variables.
I find it sensible to move these views into the stack, in order to have a single mechanism to select which view to present.

This commit is an attempt to start moving both into the stack, but I'm starting with "No Pet Selected".

As a stack page, "No Pet Selected" should be the first view `content` attempts to show.
This is in order to avoid inconsistencies when the user launches the app for the first time: the app would tried to display a pet's details, but there would be no pet.
That's why I change it from "Pet Details" to "No Pet Selected".
(Currently things work because initially `selected_pet` is `None`, but this strategy won't work for the stack.)