Remove "Edit" button from "Add Pet"
The logic to decide what to show in the header bar was a mess.
It was a single header bar and the buttons were shown based on the current view and on what the user clicked.

I moved the header bar into each stack page as an attempt to make them separate.
Now, the approach is that each stack page has its own header bar, not a shared one.
This made defining what each header bar should be way simpler (so much that I fixed them).