Extract "No Pet Selected" view into its own file
This view is defined explicitly in the stack because it's really small and because it made simpler the refactoring commit that moved it there.

I intend on putting more things there.
As such, it's sensible to extract it to avoid polluting `content`.

Extracting it also makes the stack pages declarations follow the same pattern.