Fix "Add Pet" title
When in "Add Pet", the header bar still said "Pet Details".
It should be "Add Pet" instead.

I moved "Add Pet" into the stack and it solved the issue.
This was the last view outside the stack.

Now all 4 views are part of the stack:
- "No Pet Selected"
- "Add Pet"
- "Pet Details"
- "Edit Pet"

Managing all of them is now simpler.

NOTE: The `set_visible_child` assignment was slided down to prevent warnings about attempting to present a view not yet added to the stack:
      "Child name 'No Pet Selected' not found in GtkStack"
      So, I just set the visible view _after_ adding them.