Require confirmation before deleting a pet
It just deleted the entry before any sort of confirmation.
If you pressed it accidentally, you were screwed :D

Now it shows a dialog asking to confirm the deletion or to cancel it.

I needed to use `Propagation::Stop` to avoid destroying the dialog widget, which is stored in the model.
When attempting to show it again, it would be in an inconsistent state, trying to show a destroyed widget.

Also, `connect_response` is at the bottom because I couldn't figure out a way to use the `signal[_]` syntax to clone `sender` and to pass additional arguments to it.
So I just don't use it, and because of that, I don't clone `sender`.
That's why it's at the bottom, otherwise sender would be used after this *move*.