Edit pets
Currently, the "Edit Pet" view is a placeholder, with only the "Delete" button.

This change makes it possible to actually edit pets.
It does this by adding fields to the "Edit Pet" view.

You can intepret tihs change in any of the following ways:
- The user will be able to edit whatever they put in the "Add Pet" view.
- The user will be able to edit whatever they see in "Pet Details" view.
  (Apart from the picture.)

NOTE: If a pet's name changes, the sidebar won't update until relaunch.