Apply changes in `edit_view` instead of in `content`
Currently, the changes made to a pet from "Edit Pet" is sent to `content`, and that module updates the original pet.

This change removes this delegation and makes the update from within `edit_view` itself.
In other words, the update responsibility is moved from `content` to `edit_view` in order to simplify things.