Rename `::SendPetBack(_)` to `::AddPetAndShowTheirDetails(_)`
This is for `content::Input::SendPetBack(_)`.

`SendPetBack` is a really weird name.
I named it like this because I wanted to convey the idea that I was operating on a pet and sending the result back to the parent widget.
I think `AddPet(_)` would have been a better name, as it's the naming scheme I'm using now.
However, it does more than that, as it shows the pet's details too.
Hence this new name.

It's a big name, I don't like it much, but I think it's more consistent than the old one.