Don't use `::SetVisiblePane(_)` outside of `content`
When exiting the "Edit Pet" view, it tells what `content` should show next.
"Show this pane", "show that pane" etc.
I'm not liking this approach because it requires "Edit Pet" to know what was shown before it, and what will be next.
These are not responsibilities of this view.
This is what `content` is for.
Basically it's a problem of "don't say the _how_, say the _what_".

All of that guessing game aside, this is how it works now:
- Output `::DismissPane` when clicking "Cancel".
- Only output `::DeletePet` when clicking "Delete".
  `content` decides what to show next.

This makes things simpler and clearer.