Hide empty sidebar
When the user first launches the app, they see an empty sidebar and an empty main pane.
Actually, they see `adw::StatusPage`s on both.
I find a status page in the sidebar to be weird, considering I could just hide it.

In this change I do just that: I hide the sidebar when it is empty and show it otherwise.
As a result, "No Pets Yet" is what the user first sees.
The "No Pet Selected" view is now only shown after deleting a pet (as long as the sidebar isn't empty).

One way to visualize this change is deciding which side of the split view to show as a full view.
It first shows "No Pets Yet" (sidebar side), then shows "Add Pet" (main side).
After adding a pet, it splits the view.
That's what `set_show_content: _` is for.

In simpler words, the flow is now:
- "No Pets Yet" > "Add Pet" > Sidebar + "Pet Details"

NOTES:
- From "No Pets Yet", you can only go to "Add Pet".
- When the user deletes all pets, it's the same case as launching the app for the first time.