Fix: Show "Add Pet" on single click
The first time launching the app, that is, you have no pets and the sidebar is collapsed, you need to click the "+" button twice for the "Add Pet" view to appear.
Subsequent times required only a single click.

My assumption is that it has to do with deciding which side of a split view to show as a full view.
It starts showing the sidebar side, the "No Pets Yet" view, then it should show "Add Pet", which is in the content side.
To decide whether to switch sides or not, I check `model.content.model().visible_pane` against `Panes::AddPet`.
However, I think `content` is only updated **after** `model`, so the view wouldn't see the updated value `::AddPet`.
The next time you click "+", it would check again and this time it would coincidentally have the correct value (from the previous update).
That's how I think it works, at least.

Sending a dummy message from `content` to `app` forces it to see the correct value, requiring only a single click _always_.
This solves the issue, but ideally it would just send this message when the sidebar is collapsed, but I couldn't think of anything worth the trouble of implementing.
The solutions I though of required messaging between these components anyway.
Not worth the trouble.
Better to have this simple workaround.